数据格式

+++ RKReader/Controllers/Read/RKReadPageViewController.m
@@ -148,7 +148,7 @@ RKReadMenuViewDelegate
 		self.pageNext--;
 	}
 	
-	RKLog(@"chapter:%ld -- page:%ld",self.chapterNext,self.pageNext);
+	RKLog(@"chapter:%ld -- page:%ld",(long)self.chapterNext,(long)self.pageNext);
     return [self viewControllerChapter:self.chapterNext andPage:self.pageNext];
 }
 
@@ -166,7 +166,7 @@ RKReadMenuViewDelegate
 	}else {
 		self.pageNext ++;
 	}
-	RKLog(@"chapter:%ld -- page:%ld",self.chapterNext,self.pageNext);
+	RKLog(@"chapter:%ld -- page:%ld",(long)self.chapterNext,(long)self.pageNext);
 	return [self viewControllerChapter:self.chapterNext andPage:self.pageNext];
 }
 
@@ -191,7 +191,7 @@ RKReadMenuViewDelegate
     RKLog(@"willTransitionToViewControllers");
 //	self.currentChapter = self.chapterNext;
 //	self.currentPage = self.pageNext;
-	RKLog(@"%ld -|- %ld",self.currentChapter,self.currentPage);
+	RKLog(@"%ld -|- %ld",(long)self.currentChapter,(long)self.currentPage);
 }
 
 #pragma mark - RKReadMenuViewDelegate
