+++ RKReader/Views/RKBottomStatusBar.m
@@ -45,7 +45,7 @@
 - (void)drawRect:(CGRect)rect {
 	
 	NSInteger batteryLevel = [self getCurrentBatteryLevel];
-	self.batteryNum.text = [NSString stringWithFormat:@"%ld%%",batteryLevel];
+	self.batteryNum.text = [NSString stringWithFormat:@"%ld%%",(long)batteryLevel];
 	
 	[self.batteryNum sizeToFit];
 	self.batteryNum.centerY = self.batteryImage.centerY = self.height/2;
@@ -113,7 +113,7 @@
 - (void)setBook:(RKHomeListBooks *)book {
 	_book = book;
 
-	self.bookName.text = [NSString stringWithFormat:@"%@(%ld/%ld)",self.book.fileInfo.fileName,self.book.readProgress.page+1,self.chapter.pageCount];
+	self.bookName.text = [NSString stringWithFormat:@"%@(%ld/%ld)",self.book.fileInfo.fileName,(long)self.book.readProgress.page+1,(long)self.chapter.pageCount];
 	if (self.chapters == 1) {
 		self.progress.text = [NSString stringWithFormat:@"%.2f%%",self.book.readProgress.page*1.0f/self.chapter.pageCount];
 	}else if (self.chapters == 0) {
